Moved randomReplica to internal proxy package
